serving_pipeline/requirements.txt
aiofiles==24.1.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
appdirs==1.4.4
asttokens==3.0.1
attrs==25.4.0
bigtree==1.1.0
blinker==1.9.0
boto3==1.42.19
botocore==1.42.19
brotli==1.2.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dask==2025.12.0
databricks-sdk==0.76.0
debugpy==1.8.19
decorator==5.2.1
deprecation==2.1.0
dill==0.3.9
distro==1.9.0
docker==7.1.0
dynaconf==3.2.12
evidently==0.4.33
executing==2.2.1
Faker==40.1.0
fastapi==0.128.0
feast==0.58.0
ffmpy==1.0.0
filelock==3.20.2
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.61.1
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.45.0
googleapis-common-protos==1.72.0
gradio==6.2.0
gradio_client==2.0.2
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.3.0
groovy==0.1.2
grpcio==1.76.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huey==2.5.5
huggingface_hub==1.2.3
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
kagglehub==0.3.13
kiwisolver==1.4.9
librt==0.7.7
litestar==2.19.0
litestar-htmx==0.5.0
llvmlite==0.46.0
locket==1.0.0
loguru==0.7.3
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mlflow==3.8.1
mlflow-skinny==3.8.1
mlflow-tracing==3.8.1
mmh3==5.2.0
msgspec==0.20.0
multidict==6.7.0
multipart==1.3.0
mypy==1.19.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
nltk==3.9.2
numba==0.63.1
numpy==2.3.5
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.28.9
nvitop==1.6.1
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
partd==1.4.2
pathspec==1.0.2
patsy==1.0.2
pexpect==4.9.0
pillow==12.0.0
pip==25.3
platformdirs==4.5.1
plotly==5.24.1
pluggy==1.6.0
polyfactory==3.2.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==6.33.2
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
pykernel==0.1.6
pyparsing==3.3.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
redis==7.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rich-click==1.9.5
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
safehttpx==0.1.7
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
semantic-version==2.10.0
setuptools==80.9.0
shap==0.49.1
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.45
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.50.0
statsmodels==0.14.6
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.6.0
toml==0.10.2
tomlkit==0.13.3
toolz==1.1.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typeguard==4.4.4
typer==0.21.0
typer-slim==0.21.0
typing_extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.2
uuid6==2025.0.1
uvicorn==0.34.0
uvicorn-worker==0.3.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.4
wheel==0.45.1
xgboost==3.1.2
zipp==3.23.0
